spreadsheets/requirements.txt
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
google-auth==1.14.0
gspread==3.4.2
idna==2.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.23.0
rsa==4.0
six==1.14.0
urllib3==1.25.9
